v1.4: platform-app on Postgres + Redis (drop sqlite + in-memory)
Single-shot rewrite — no dual backend, no sqlite fallback. State lives
entirely in Postgres + Redis so the application container is fully
stateless and can run on Railway / DO / wherever.

Changes:
- pyproject.toml: + psycopg[binary], + redis; bump version 1.4.0
- settings.py: required env DATABASE_URL + REDIS_URL (fail-fast on boot);
  dropped db_path + proxy_log_db_path
- db.py: full rewrite around psycopg.Connection. _DB wrapper auto-
  reconnects on broken connection so the existing
  `db.main().execute(sql, params).fetchone()` API stays the same. Both
  main and proxy_log share one Postgres now (was two sqlite files).
- cache.py: new module — Redis-backed _Cache replaces TTLCache for
  tenant/session/acl. Multi-instance safe.
- hmac_sign.py: NonceStore now uses Redis SETNX (atomic check-and-add
  with TTL); gc() is a no-op (Redis auto-expires).
- migrations/001_init.sql: Postgres syntax (BIGINT for timestamps,
  TEXT/INTEGER kept).
- migrations/002_proxy_log.sql: Postgres BIGSERIAL for autoincrement.
- auth.py / admin.py / api.py / health.py: ? -> %s placeholders;
  admin.cmd_grant uses ON CONFLICT ... DO UPDATE (postgres equivalent
  of sqlite INSERT OR REPLACE).
- tests/conftest.py: env vars before import; TRUNCATE all tables +
  flushdb redis between tests.
- tests/test_db.py: rewrite (sqlite_master -> information_schema;
  proxy_log writability test instead of "separate file").
- tests/test_hmac_sign.py::test_nonce_store_gc: rewritten as
  test_nonce_store_redis_expiry — gc() is no-op now.
- Dockerfile: CMD uses ${PORT:-80} so Railway can inject $PORT.

27 tests pass against local Postgres 16 + Redis 7 containers.
uvicorn smoke: GET / -> 200 (login.html), GET /api/me -> 401.

NOT in this commit (separate work): yinhu agent rewrite,
data migration script, Railway deployment config.

+++ platform/migrations/001_init.sql
@@ -4,8 +4,8 @@ CREATE TABLE IF NOT EXISTS users (
   password_hash TEXT NOT NULL,
   display_name  TEXT NOT NULL,
   email         TEXT,
-  created_at    INTEGER NOT NULL,
-  last_login    INTEGER
+  created_at    BIGINT NOT NULL,
+  last_login    BIGINT
 );
 
 CREATE TABLE IF NOT EXISTS tenants (
@@ -17,17 +17,17 @@ CREATE TABLE IF NOT EXISTS tenants (
   hmac_key_id_current      TEXT NOT NULL,
   hmac_secret_prev         TEXT NOT NULL DEFAULT '',
   hmac_key_id_prev         TEXT NOT NULL DEFAULT '',
-  hmac_rotated_at          INTEGER,
+  hmac_rotated_at          BIGINT,
   agent_version            TEXT NOT NULL DEFAULT 'unknown',
   health                   TEXT NOT NULL DEFAULT 'unknown',
-  health_checked_at        INTEGER,
+  health_checked_at        BIGINT,
   allowed_response_headers TEXT NOT NULL DEFAULT '[]',
   icon_url                 TEXT,
   description              TEXT,
   visibility               TEXT NOT NULL DEFAULT 'private',
   active                   INTEGER NOT NULL DEFAULT 1,
   tenant_uid               TEXT NOT NULL UNIQUE,
-  created_at               INTEGER NOT NULL,
+  created_at               BIGINT NOT NULL,
   PRIMARY KEY (client_id, agent_id)
 );
 
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS user_tenant (
   client_id  TEXT NOT NULL,
   agent_id   TEXT NOT NULL,
   role       TEXT NOT NULL DEFAULT 'user',
-  granted_at INTEGER NOT NULL,
+  granted_at BIGINT NOT NULL,
   granted_by TEXT,
   PRIMARY KEY (user_id, client_id, agent_id),
   FOREIGN KEY (client_id, agent_id) REFERENCES tenants(client_id, agent_id)
@@ -46,8 +46,8 @@ CREATE TABLE IF NOT EXISTS platform_sessions (
   id         TEXT PRIMARY KEY,
   user_id    TEXT NOT NULL REFERENCES users(id),
   csrf_token TEXT NOT NULL,
-  created_at INTEGER NOT NULL,
-  expires_at INTEGER NOT NULL,
+  created_at BIGINT NOT NULL,
+  expires_at BIGINT NOT NULL,
   ip         TEXT,
   user_agent TEXT
 );
@@ -62,10 +62,10 @@ CREATE TABLE IF NOT EXISTS api_keys (
   hash              TEXT NOT NULL,
   scope             TEXT NOT NULL DEFAULT 'rw',
   source_session_id TEXT REFERENCES platform_sessions(id),
-  expires_at        INTEGER,
-  created_at        INTEGER NOT NULL,
-  last_used         INTEGER,
-  revoked_at        INTEGER,
+  expires_at        BIGINT,
+  created_at        BIGINT NOT NULL,
+  last_used         BIGINT,
+  revoked_at        BIGINT,
   FOREIGN KEY (client_id, agent_id) REFERENCES tenants(client_id, agent_id)
 );
